refactor database to be more readable implement get request for product by id
