+++ database/seed.js
@@ -1,9 +1,13 @@
 const mongoose = require('mongoose');
+const dbConnect = require('./dbConnect.js');
 const Product = require('./Product.js');
-const products = require('./data.js');
+const products = require('./data/data.js');
 
 // set up database
-mongoose.connect('mongodb://localhost/fjords', { useNewUrlParser: true, useUnifiedTopology: true });
+// mongoose.connect('mongodb://localhost/fjords', { useNewUrlParser: true, useUnifiedTopology: true });
+
+// create and connect to database
+dbConnect();
 
 Product.create(products)
   .then(() => mongoose.disconnect())
